add voters & fixes

+++ AllWeShare/src/Controller/Front/PostController.php
@@ -63,7 +63,7 @@ class PostController extends AbstractController
         return $this->render('Front/post/show.html.twig', [
             'post' => $post,
             'comment' => $comment,
-            'comments' => $commentRepository->findBy(['post' => $post]),
+            'comments' => $commentRepository->findBy(['post' => $post], ['createdAt' => 'ASC']),
             'form' => $form->createView(),
             ]);
     }
@@ -73,6 +73,7 @@ class PostController extends AbstractController
      */
     public function edit(Request $request, Post $post): Response
     {
+        $this->denyAccessUnlessGranted('edit', $post);
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
 
@@ -93,6 +94,7 @@ class PostController extends AbstractController
      */
     public function delete(Request $request, Post $post): Response
     {
+        $this->denyAccessUnlessGranted('delete', $post);
         if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($post);
